extracting pipeline logic  Please enter the commit message for your changes. Lines starting
